Fixed aggregation sorting on equal values
